# internal/pkg/api/userService_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: userService.proto

package api

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationUserDelUserDeptAssoc = "/api.User/DelUserDeptAssoc"
const OperationUserEditUserName = "/api.User/EditUserName"
const OperationUserGetUserPermissions = "/api.User/GetUserPermissions"
const OperationUserLogin = "/api.User/Login"

type UserHTTPServer interface {
	// DelUserDeptAssoc 从部门中移除用户
	DelUserDeptAssoc(context.Context, *DelUserDeptAssocRequest) (*Empty, error)
	// EditUserName 修改用户名
	EditUserName(context.Context, *EditUserNameRequest) (*Empty, error)
	// GetUserPermissions --------------------------------------------------
	// 查询用户在某项目内所有权限，去掉角色这一层，直接列出所有拥有的权限值
	GetUserPermissions(context.Context, *GetUserPermissionsRequest) (*GetUserPermissionsResponse, error)
	// Login 登录或注册，其实可以理解为只是通过用户账号密码新建一个token，用于其他接口鉴权
	Login(context.Context, *LoginRequest) (*LoginResponse, error)
}

func RegisterUserHTTPServer(s *http.Server, srv UserHTTPServer) {
	r := s.Route("/")
	r.POST("/user/token", _User_Login0_HTTP_Handler(srv))
	r.PATCH("/user", _User_EditUserName0_HTTP_Handler(srv))
	r.DELETE("/user-dept-assoc", _User_DelUserDeptAssoc0_HTTP_Handler(srv))
	r.GET("/user/permissions", _User_GetUserPermissions0_HTTP_Handler(srv))
}

func _User_Login0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserLogin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Login(ctx, req.(*LoginRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginResponse)
		return ctx.Result(200, reply)
	}
}

func _User_EditUserName0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditUserNameRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserEditUserName)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditUserName(ctx, req.(*EditUserNameRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Empty)
		return ctx.Result(200, reply)
	}
}

func _User_DelUserDeptAssoc0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DelUserDeptAssocRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserDelUserDeptAssoc)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DelUserDeptAssoc(ctx, req.(*DelUserDeptAssocRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Empty)
		return ctx.Result(200, reply)
	}
}

func _User_GetUserPermissions0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetUserPermissionsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserGetUserPermissions)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUserPermissions(ctx, req.(*GetUserPermissionsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetUserPermissionsResponse)
		return ctx.Result(200, reply)
	}
}

type UserHTTPClient interface {
	// DelUserDeptAssoc 从部门中移除用户
	DelUserDeptAssoc(ctx context.Context, req *DelUserDeptAssocRequest, opts ...http.CallOption) (rsp *Empty, err error)
	// EditUserName 修改用户名
	EditUserName(ctx context.Context, req *EditUserNameRequest, opts ...http.CallOption) (rsp *Empty, err error)
	// GetUserPermissions --------------------------------------------------
	// 查询用户在某项目内所有权限，去掉角色这一层，直接列出所有拥有的权限值
	GetUserPermissions(ctx context.Context, req *GetUserPermissionsRequest, opts ...http.CallOption) (rsp *GetUserPermissionsResponse, err error)
	// Login 登录或注册，其实可以理解为只是通过用户账号密码新建一个token，用于其他接口鉴权
	Login(ctx context.Context, req *LoginRequest, opts ...http.CallOption) (rsp *LoginResponse, err error)
}

type UserHTTPClientImpl struct {
	cc *http.Client
}

func NewUserHTTPClient(client *http.Client) UserHTTPClient {
	return &UserHTTPClientImpl{client}
}

// DelUserDeptAssoc 从部门中移除用户
func (c *UserHTTPClientImpl) DelUserDeptAssoc(ctx context.Context, in *DelUserDeptAssocRequest, opts ...http.CallOption) (*Empty, error) {
	var out Empty
	pattern := "/user-dept-assoc"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserDelUserDeptAssoc))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// EditUserName 修改用户名
func (c *UserHTTPClientImpl) EditUserName(ctx context.Context, in *EditUserNameRequest, opts ...http.CallOption) (*Empty, error) {
	var out Empty
	pattern := "/user"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserEditUserName))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetUserPermissions --------------------------------------------------
// 查询用户在某项目内所有权限，去掉角色这一层，直接列出所有拥有的权限值
func (c *UserHTTPClientImpl) GetUserPermissions(ctx context.Context, in *GetUserPermissionsRequest, opts ...http.CallOption) (*GetUserPermissionsResponse, error) {
	var out GetUserPermissionsResponse
	pattern := "/user/permissions"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserGetUserPermissions))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Login 登录或注册，其实可以理解为只是通过用户账号密码新建一个token，用于其他接口鉴权
func (c *UserHTTPClientImpl) Login(ctx context.Context, in *LoginRequest, opts ...http.CallOption) (*LoginResponse, error) {
	var out LoginResponse
	pattern := "/user/token"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserLogin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}