# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /user:
        get:
            tags:
                - UserCURD
            operationId: UserCURD_GetUserList
            parameters:
                - name: userNameList
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetUserListResponse'
        post:
            tags:
                - UserCURD
            description: |-
                --------------------------------------------------
                 tbl : user
            operationId: UserCURD_AddUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.AddUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.AddUserResponse'
        delete:
            tags:
                - UserCURD
            operationId: UserCURD_DelUserByIDList
            parameters:
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.Empty'
        patch:
            tags:
                - User
            description: 修改用户名
            operationId: User_EditUserName
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.EditUserNameRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.Empty'
...
    /user-dept-assoc:
        get:
            tags:
                - UserCURD
            operationId: UserCURD_GetUserDeptAssocList
            parameters:
                - name: userIDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
                - name: deptIDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetUserDeptAssocListResponse'
        post:
            tags:
                - UserCURD
            description: |-
                --------------------------------------------------
                 tbl : user_dept_assoc
            operationId: UserCURD_AddUserDeptAssoc
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.AddUserDeptAssocRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.AddUserDeptAssocResponse'
        delete:
            tags:
                - User
            description: 从部门中移除用户
            operationId: User_DelUserDeptAssoc
            parameters:
                - name: userID
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: deptID
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.Empty'
        patch:
            tags:
                - UserCURD
            operationId: UserCURD_UpdateUserDeptAssoc
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.UpdateUserDeptAssocRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.UpdateUserDeptAssocResponse'
...
    /user/permissions:
        get:
            tags:
                - User
            description: |-
                --------------------------------------------------
                 查询用户在某项目内所有权限，去掉角色这一层，直接列出所有拥有的权限值
            operationId: User_GetUserPermissions
            parameters:
                - name: userID
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: projectID
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetUserPermissionsResponse'
    /user/token:
        post:
            tags:
                - User
            description: 登录或注册，其实可以理解为只是通过用户账号密码新建一个token，用于其他接口鉴权
            operationId: User_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.LoginResponse'
components:
    schemas:
...
        api.AddUserDeptAssocRequest:
            type: object
            properties:
                userDeptAssoc:
                    $ref: '#/components/schemas/api.UserDeptAssocInfo'
        api.AddUserDeptAssocResponse:
            type: object
            properties:
                userDeptAssoc:
                    $ref: '#/components/schemas/api.UserDeptAssocInfo'
...
        api.AddUserRequest:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/api.UserInfo'
        api.AddUserResponse:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/api.UserInfo'
...
        api.EditUserNameRequest:
            type: object
            properties:
                ID:
                    type: integer
                    format: int32
                userName:
                    type: string
        api.Empty:
            type: object
            properties: {}
...
        api.GetUserDeptAssocListResponse:
            type: object
            properties:
                userDeptAssocList:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.UserDeptAssocInfo'
...
        api.GetUserListResponse:
            type: object
            properties:
                userList:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.UserInfo'
        api.GetUserPermissionsResponse:
            type: object
            properties:
                actionToPathPattern:
                    type: object
                    additionalProperties:
                        type: string
...
        api.LoginRequest:
            type: object
            properties:
                userName:
                    type: string
        api.LoginResponse:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/api.UserInfo'
                token:
                    type: string
...
        api.UpdateUserDeptAssocRequest:
            type: object
            properties:
                userDeptAssoc:
                    $ref: '#/components/schemas/api.UserDeptAssocInfo'
        api.UpdateUserDeptAssocResponse:
            type: object
            properties:
                userDeptAssoc:
                    $ref: '#/components/schemas/api.UserDeptAssocInfo'
...
        api.UserDeptAssocInfo:
            type: object
            properties:
                ID:
                    type: integer
                    format: int32
                createTime:
                    type: string
                createUser:
                    type: integer
                    format: int32
                updateTime:
                    type: string
                updateUser:
                    type: integer
                    format: int32
                userID:
                    type: integer
                    format: int32
                deptID:
                    type: integer
                    format: int32
                jobID:
                    type: integer
                    format: int32
            description: |-
                --------------------------------------------------
                 tbl : user_dept_assoc
...
        api.UserInfo:
            type: object
            properties:
                ID:
                    type: integer
                    format: int32
                createTime:
                    type: string
                createUser:
                    type: integer
                    format: int32
                updateTime:
                    type: string
                updateUser:
                    type: integer
                    format: int32
                userName:
                    type: string
                password:
                    type: string
                lastEditFrom:
                    type: string
            description: |-
                --------------------------------------------------
                 tbl : user
...
servers:
    - description: PGO API
      url: http://127.0.0.1:8080